cast time(0) to fix gcc warning on 64 bit host

+++ C/ds18b20_to_json.c
@@ -75,7 +75,7 @@ int main(int argc, char **argv)
         return -1;
     }
 
-    printf("{\"t\":%lld, \"temp\":%.2f, \"device\":\"DS18B20\", \"DS18B20_ID\":\"%s\"}", time(0), (float)raw_val * 9 / 5000 + 32, argv[1]);
+    printf("{\"t\":%lld, \"temp\":%.2f, \"device\":\"DS18B20\", \"DS18B20_ID\":\"%s\"}", (long long)time(0), (float)raw_val * 9 / 5000 + 32, argv[1]);
 
     exit(0);
 }
